google/poppy: enable trackpad as wake source
This configures GPP_A23 as a wake source for the trackpad.  We also
need to set up GPP_A GPE0_DW0, thus evicting GPP_B.  We don't have any
interesting signals in GPP_B, so we won't be missing it.

I don't have hardware with A23 wired up, so i just tested the wake
source using A19 which is essentially identical to A23.

BUG=b:78541883
TEST=verified we can trackpad can wake system from suspend

Change-Id: If800464c8b2319d758b1823850571919f85bdc6c
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://review.coreboot.org/25850
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Furquan Shaikh <furquan@google.com>